Add New Function : Delete, Update and Select Specific User

+++ src/main/java/com/example/demo/repository/StudentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends CrudRepository<Student, Long> {
+public interface StudentRepository extends CrudRepository<Student, Integer> {
     // Your custom query methods, if any
 }
